moved widgets to own files
